A workaround for possible gfortran bug
See #9
this adds a small tol for the equality comparisons in the test. the atan() function is the only one that isn't exactly equal (on gfortran 11.1.0)

workaround for possible gfortran bug

See #9
this adds a small tol for the equality comparisons in the test. the atan() function is the only one that isn't exactly equal (on gfortran 11.1.0)